Make attestation derive only, remove crml-generic-asset and crml-attestion packages
