+++ README.md
@@ -17,6 +17,4 @@ These libraries provide some syntactic sugar and extensions to the core `@cennzn
 
 | Name                                                        | Description                                                          |
 | ----------------------------------------------------------- | -------------------------------------------------------------------- |
-| [@cennznet/crml-generic-asset](packages/crml-generic-asset) | An sdk providing additional features for generic asset runtime module |
 | [@cennznet/crml-cennzx-spot](packages/crml-cennzx-spot)     | An sdk providing additional features for CENNZX spot runtime module   |
-| [@cennznet/crml-attestation](packages/crml-attestation)     | An sdk providing additional features for CENNZnet's attestation module |
