+++ jest/integration.config.js
@@ -12,22 +12,19 @@ module.exports = {
   testEnvironment: '<rootDir>/jest/env.js',
   moduleNameMapper: {
     '@cennznet/api(.*)$': '<rootDir>/packages/api/src/$1',
-    '@cennznet/crml-attestation(.*)$': '<rootDir>/packages/crml-attestation/src/$1',
-    '@cennznet/crml-generic-asset(.*)$': '<rootDir>/packages/crml-generic-asset/src/$1',
     '@cennznet/crml-cennzx-spot(.*)$': '<rootDir>/packages/crml-cennzx-spot/src/$1',
     '@cennznet/types(.*)$': '<rootDir>/packages/types/src/$1',
     '@cennznet/util(.*)$': '<rootDir>/packages/util/src/$1',
   },
   modulePathIgnorePatterns: [
     '<rootDir>/packages/api/build',
-    '<rootDir>/packages/crml-attestation/build',
-    '<rootDir>/packages/crml-generic-asset/build',
     '<rootDir>/packages/crml-cennzx-spot/build',
     '<rootDir>/packages/types/build',
     '<rootDir>/packages/util/build',
   ],
   setupFilesAfterEnv: ['<rootDir>/jest/jest.init.js'],
-  globals: { // Added global flag to enable doughnut wasm package while running integration tests
+  globals: {
+    // Enable doughnut wasm package while running integration tests
     DOUGHNUT_SUPPORT: true
   }
   // globalSetup: '<rootDir>/jest/jest.setup.js',
