+++ package.json
@@ -21,10 +21,12 @@
   "devDependencies": {
     "@babel/cli": "^7.0.0",
     "@babel/core": "^7.7.4",
+    "@babel/helper-skip-transparent-expression-wrappers": "^7.11.0",
+    "@babel/helper-validator-identifier": "^7.10.4",
     "@babel/plugin-proposal-numeric-separator": "^7.10.4",
     "@babel/preset-env": "^7.7.6",
     "@babel/runtime-corejs3": "^7.7.6",
-    "@polkadot/dev": "^0.31.1",
+    "@polkadot/dev": "0.31.1",
     "@polkadot/ts": "^0.1.86",
     "@types/jest": "^23.3.10",
     "@typescript-eslint/eslint-plugin": "^3.10.0",
