+++ packages/api/package.json
@@ -19,9 +19,7 @@
     "doc": "npx compodoc --tsconfig ../../tsconfig.json --hideGenerator --theme readthedocs --includes compodoc --output ../../docs/api"
   },
   "dependencies": {
-    "@cennznet/crml-attestation": "^1.2.1",
     "@cennznet/crml-cennzx-spot": "^1.2.1",
-    "@cennznet/crml-generic-asset": "^1.2.1",
     "@cennznet/types": "^1.2.1",
     "@polkadot/api": "^1.1.1",
     "@polkadot/metadata": "1.1.1",
