+++ packages/api/src/derives/index.ts
@@ -13,19 +13,18 @@
 // limitations under the License.
 
 import { ApiTypes } from '@cennznet/api/types';
-import * as attestation from '@cennznet/crml-attestation/derives';
 import * as cennzxSpot from '@cennznet/crml-cennzx-spot/derives';
-import * as genericAsset from '@cennznet/crml-generic-asset/derives';
 import { AnyFunction } from '@cennznet/types/types';
 import { ApiInterfaceRx, MethodResult } from '@polkadot/api/types';
 import { Observable } from 'rxjs';
+import * as attestation from './attestation';
 import * as fees from './fees';
 import * as session from './session';
 import * as staking from './staking';
 
 export type DeriveFunc = (api: ApiInterfaceRx) => (...args: any[]) => Observable<any>;
 
-export const derive = { attestation, cennzxSpot, fees, staking, session, genericAsset };
+export const derive = { attestation, cennzxSpot, fees, staking, session };
 
 export type DecoratedCennznetDerive<
   ApiType extends ApiTypes,
