--- packages/crml-attestation/examples/index.js
@@ -1,43 +1,0 @@
-// Copyright 2019 Centrality Investments Limited
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-const { Api } = require('@cennznet/api');
-const { Attestation } = require('../dist');
-
-import testKeyring from '@polkadot/keyring/testing';
-
-const issuerUri = '//Alice';
-const holderUri = '//Charlie';
-
-async function main() {
-  let simpleKeyring = testKeyring({ type: 'sr25519' });
-  const issuer = simpleKeyring.addFromUri(issuerUri);
-  const holder = simpleKeyring.addFromUri(holderUri);
-  const api = await Api.create({
-    provider: 'wss://cennznet-node-0.centrality.me:9944',
-  });
-  const attestation = new Attestation(api);
-
-  const topic = 'test';
-  const value = '1234';
-  const claim = await attestation.setClaim(holder.address, topic, value);
-
-  await claim.signAndSend(issuer, async result => {
-    if (result.type === 'Finalised' && result.events !== undefined) {
-      const { data } = result.events[0].event.toJSON();
-      console.log(data);
-    }
-  });
-}
-main();
