--- packages/crml-attestation/package.json
@@ -1,26 +1,0 @@
-{
-    "name": "@cennznet/crml-attestation",
-    "version": "1.2.1",
-    "description": "cennznet's javascript library for supporting identity runtime",
-    "author": "Centrality Development Team",
-    "homepage": "https://github.com/cennznet/api.js/tree/master/packages/crml-attestation",
-    "license": "Apache-2.0",
-    "main": "index.js",
-    "types": "index.d.ts",
-    "repository": {
-        "type": "git",
-        "url": "git+ssh://git@github.com/cennznet/api.js.git"
-    },
-    "scripts": {},
-    "dependencies": {
-        "@cennznet/api": "^1.2.1",
-        "@cennznet/types": "^1.2.1",
-        "@cennznet/util": "^1.2.1"
-    },
-    "devDependencies": {
-        "@polkadot/keyring": "2.3.1",
-        "@types/jest": "^24.0.15",
-        "jest": "24.8.0",
-        "rimraf": "^2.6.2"
-    }
-}
