--- packages/crml-generic-asset/CHANGELOG.md
@@ -1,51 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [0.20.0](https://github.com/cennznet/crml-generic-asset.js) (2019-09-10)
-* Upgrade to `@plugnet/api@0.90.x`
-
-
-### Breaking Change
-* GenericAsset.create() and GenericAssetRx.create() are removed, they are not meant to be instantiated manually.
-
-
-## [0.8.1](https://github.com/cennznet/crml-generic-asset.js/compare/v0.8.0...v0.8.1) (2019-05-21)
-
-
-### Bug Fixes
-
-* revert change of AssetBalance ([d6c817b](https://github.com/cennznet/crml-generic-asset.js/commit/d6c817b)), closes [#27](https://github.com/cennznet/crml-generic-asset.js/issues/27)
-
-
-### Features
-
-* queries return as AssetBalance which support unit conversions ([0ab5f76](https://github.com/cennznet/crml-generic-asset.js/commit/0ab5f76)), closes [#26](https://github.com/cennznet/crml-generic-asset.js/issues/26)
-
-
-
-
-
-# [0.8.0](https://bitbucket.org/centralitydev/cennznet-js-generic-asset/compare/v0.7.3...v0.8.0) (2019-05-13)
-
-
-### Dependency update
-
-* update to polkadot 0.13.1 ([c34648a](https://bitbucket.org/centralitydev/cennznet-js-generic-asset/commits/c34648a)), closes [#25](https://bitbucket.org/centralitydev/cennznet-js-generic-asset/issue/25)
-
-
-
-
-
-## [0.7.3](https://bitbucket.org/centralitydev/cennznet-js-generic-asset/compare/v0.7.1...v0.7.3) (2019-05-09)
-
-**Note:** Version bump only for package @cennznet/crml-generic-asset
-
-
-
-
-
-## [0.7.1](https://bitbucket.org/centralitydev/cennznet-js-generic-asset/compare/v0.7.0...v0.7.1) (2019-05-06)
-
-**Note:** Version bump only for package @cennznet/crml-generic-asset
