--- packages/crml-generic-asset/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@cennznet/crml-generic-asset",
-  "version": "1.2.1",
-  "description": "the javascript sdk for cennznet runtime module: Generic Asset. also a plugin of @cennznet/api",
-  "author": "Alex Wang <alex.wang@centrality.ai>",
-  "maintainers": [
-    "Centrality Development Team"
-  ],
-  "homepage": "https://github.com/cennznet/crml-generic-asset.js",
-  "license": "Apache-2.0",
-  "main": "index.js",
-  "types": "index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:cennznet/crml-generic-asset.js.git"
-  },
-  "scripts": {},
-  "dependencies": {
-    "@cennznet/api": "^1.2.1",
-    "@cennznet/types": "^1.2.1",
-    "@cennznet/util": "^1.2.1"
-  },
-  "devDependencies": {
-    "@polkadot/keyring": "2.3.1",
-    "@types/jest": "^24.0.15",
-    "jest": "24.8.0"
-  }
-}
