--- packages/crml-generic-asset/src/utils/utils.ts
@@ -1,23 +1,0 @@
-// Copyright 2019 Centrality Investments Limited
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import { isBn, isNumber, isObject, isU8a } from '@cennznet/util';
-
-export function isAssetObj(value: any): boolean {
-  return isObject(value) && !isU8a(value) && !isBn(value) && isNumber(value.id);
-}
-
-export function isStringNumber(str: string): boolean {
-  return !isNaN(Number(str));
-}
