+++ scripts/generateSlimMetadataWrapper.js
@@ -4,19 +4,17 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 require('@babel/register')({
-    extensions: ['.js', '.ts'],
-    plugins: [
-        ['module-resolver', {
-            alias: {
-                '^@cennznet/api(.*)': './packages/api/src\\1',
-                '^@cennznet/types(.*)': './packages/types/src\\1',
-                '^@cennznet/crml-attestation(.*)': './packages/crml-attestation/src\\1',
-                '^@cennznet/crml-cennzx-spot(.*)': './packages/crml-cennzx-spot/src\\1',
-                '^@cennznet/crml-generic-asset(.*)': './packages/crml-generic-asset/src\\1',
-                '^@cennznet/util(.*)': './packages/util/src\\1',
-            }
-        }]
-    ]
+  extensions: ['.js', '.ts'],
+  plugins: [
+    ['module-resolver', {
+      alias: {
+        '^@cennznet/api(.*)': './packages/api/src\\1',
+        '^@cennznet/types(.*)': './packages/types/src\\1',
+        '^@cennznet/crml-cennzx-spot(.*)': './packages/crml-cennzx-spot/src\\1',
+        '^@cennznet/util(.*)': './packages/util/src\\1',
+      }
+    }]
+  ]
 });
 
 require('./generateSlimMetadata.ts');
